fix: regressions in api-tracker server
